Relationships

+++ app/Http/Controllers/AppointmentController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use App\Models\Appointment;
 use App\Http\Resources\Appointment as AppointmentResource;
 use App\Http\Resources\AppointmentCollection;
+use Illuminate\Support\Facades\Auth;
 
 class AppointmentController extends Controller
 {
@@ -28,16 +29,22 @@ class AppointmentController extends Controller
 
     public function index()
     {
+        $this->authorize('viewAny', Appointment::class);
+
         return new AppointmentCollection(Appointment::paginate(10));
     }
 
     public function show(Appointment $appointment)
     {
+        $this->authorize('view', $appointment);
+
         return response()->json(new AppointmentResource($appointment), 200);
     }
 
     public function store(Request $request)
     {
+        $this->authorize('create', Appointment::class);
+
         $request->validate(self::$rules, self::$messages);
         $appointment = new Appointment($request->all());
         $appointment->save();
@@ -47,6 +54,8 @@ class AppointmentController extends Controller
 
     public function update(Request $request, Appointment $appointment)
     {
+        $this->authorize('update', $appointment);
+
         $request->validate(self::$repulses, self::$messages);
         $appointment->update($request->all());
         return response()->json($appointment, 200);
@@ -54,7 +63,23 @@ class AppointmentController extends Controller
 
     public function delete(Request $request, Appointment $appointment)
     {
+        $this->authorize('delete', $appointment);
+
         $appointment->delete();
         return response()->json(null, 204);
+    }
+
+    public function indexPartner()
+    {
+        $this->authorize('view', Appointment::class);
+        $user = Auth::user();
+        return new AppointmentCollection($user->appointments_partner);
+    }
+
+    public function indexAfiliate()
+    {
+        $this->authorize('view', Appointment::class);
+        $user = Auth::user();
+        return new AppointmentCollection($user->appointments_afiliate);
     }
 }
